backend/requirements.txt
fastapi
uvicorn
pillow
openpyxl
